# docs/swagger.yaml
basePath: /api
definitions:
  structs.AktivasiUserResponse:
    properties:
      data:
        $ref: '#/definitions/structs.UserObat'
      pesan:
        example: Aktivasi user berhasil
        type: string
      status:
        example: sukses
        type: string
    type: object
...
  structs.DeleteUserResponse:
    properties:
      pesan:
        example: User berhasil dihapus
        type: string
      status:
        example: sukses
        type: string
    type: object
...
  structs.ErrorResponse:
    properties:
      pesan:
        example: Terjadi kesalahan
        type: string
      status:
        example: error
        type: string
    type: object
...
  structs.GetUserByIDResponse:
    properties:
      data:
        $ref: '#/definitions/structs.UserObat'
      pesan:
        example: Get user by ID berhasil
        type: string
      status:
        example: sukses
        type: string
    type: object
  structs.GetUserResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/structs.UserObat'
        type: array
      pesan:
        example: Get user berhasil
        type: string
      status:
        example: sukses
        type: string
    type: object
...
  structs.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  structs.LoginResponse:
    properties:
      pesan:
        example: 'Hi, dodo Role: ADMINFARMASI'
        type: string
      status:
        example: sukses
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  structs.RegisterUserResponse:
    properties:
      data:
        $ref: '#/definitions/structs.UserObat'
      pesan:
        example: registrasi user berhasil
        type: string
      status:
        example: sukses
        type: string
    type: object
...
  structs.TambahUserRequest:
    properties:
      nama:
        type: string
      password:
        minLength: 6
        type: string
      role:
        enum:
        - KEPALAFARMASI
        - ADMINFARMASI
        - APOTEKER
        type: string
      username:
        type: string
    required:
    - nama
    - password
    - role
    - username
    type: object
  structs.TambahUserResponse:
    properties:
      data:
        $ref: '#/definitions/structs.UserObat'
      pesan:
        example: User berhasil ditambahkan
        type: string
      status:
        example: sukses
        type: string
    type: object
...
  structs.UpdateUserRequest:
    properties:
      nama:
        type: string
      password:
        type: string
      role:
        enum:
        - KEPALAFARMASI
        - ADMINFARMASI
        - APOTEKER
        type: string
    type: object
  structs.UpdateUserResponse:
    properties:
      data:
        $ref: '#/definitions/structs.UserObat'
      pesan:
        example: Update user berhasil
        type: string
      status:
        example: sukses
        type: string
    type: object
  structs.UserObat:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: integer
      modified_at:
        type: string
      modified_by:
        type: string
      nama:
        type: string
      role:
        type: string
      status:
        type: boolean
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: '-'
    name: Damar
  description: 'Dokumentasi REST API Alur Obat Farmasi menggunakan Golang (Framework
    : Gin dan Database : PostgreSQL).'
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Final Project API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login menggunakan username dan password untuk mendapatkan JWT Token
      parameters:
      - description: Data Login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login berhasil
          schema:
            $ref: '#/definitions/structs.LoginResponse'
        "400":
          description: Request login tidak valid
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Username atau password salah
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Login User
      tags:
      - 1. Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Menambahkan user baru
      parameters:
      - description: Registrasi Data User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.TambahUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Registrasi user berhasil
          schema:
            $ref: '#/definitions/structs.RegisterUserResponse'
        "400":
          description: Request Register tidak valid
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Register User
      tags:
      - 1. Authentication
  /kepala-farmasi/users:
    get:
      description: Mengambil seluruh data user
      produces:
      - application/json
      responses:
        "200":
          description: Get User
          schema:
            $ref: '#/definitions/structs.GetUserResponse'
        "401":
          description: Gagal get User
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Semua User
      tags:
      - 2. User
    post:
      consumes:
      - application/json
      description: Menambahkan user baru
      parameters:
      - description: Data User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.TambahUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Tambah user berhasil
          schema:
            $ref: '#/definitions/structs.TambahUserResponse'
        "400":
          description: Gagal tambah user
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tambah User
      tags:
      - 2. User
  /kepala-farmasi/users/{id}:
    delete:
      description: Menghapus user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: delete user
          schema:
            $ref: '#/definitions/structs.DeleteUserResponse'
        "400":
          description: ID User tidak valid
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: User tidak ditemukan
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete User
      tags:
      - 2. User
    get:
      description: Mengambil detail user berdasarkan ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Get User by ID
          schema:
            $ref: '#/definitions/structs.GetUserByIDResponse'
        "400":
          description: ID User tidak valid
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Gagal get User by ID
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get User By ID
      tags:
      - 2. User
    put:
      consumes:
      - application/json
      description: Mengubah data user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Data User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Update User
          schema:
            $ref: '#/definitions/structs.UpdateUserResponse'
        "400":
          description: Gagal update User
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: User tidak ditemukan
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update User
      tags:
      - 2. User
  /kepala-farmasi/users/{id}/aktivasi:
    patch:
      consumes:
      - application/json
      description: Mengaktifkan atau menonaktifkan user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.AktivasiUserResponse'
      produces:
      - application/json
      responses:
        "200":
          description: aktivasi user
          schema:
            $ref: '#/definitions/structs.AktivasiUserResponse'
        "400":
          description: gagal aktivasi user
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: ID user tidak ditemukan
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Aktivasi User
      tags:
      - 2. User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...